🛠️ refactor: extract terminal footer pickers and RPC channels into dedicated modules and remove unused unsupported agent.
